requirements.txt
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
torchinfo==1.8.0
torchsummary==1.5.1
tqdm==4.66.4